Add route for automatic redirection of incoming log urls
